+++ package.json
@@ -32,8 +32,7 @@
     "mongodb-runner": "5.9.2",
     "nyc": "17.1.0",
     "parse-server": "8.2.1",
-    "semantic-release": "24.2.5",
-    "sinon": "21.0.0"
+    "semantic-release": "24.2.5"
   },
   "scripts": {
     "lint": "eslint --cache ./index.js ./lib/** ./spec/**/*.js",
